+++ patterns/behavioral-design-patterns/command/command.interface.ts
@@ -1,0 +1,11 @@
+/**
+ * @author       Digitsensitive <digit.sensitivee@gmail.com>
+ * @copyright    2019 Digitsensitive
+ * @description  Command Interface
+ *               Declares a single execution method
+ * @license      Digitsensitive
+ */
+
+export interface Command {
+  execute(): void;
+}